feat: decouple `fvm install` from self install
