+++ crates/fvm/src/bin/command/mod.rs
@@ -1,2 +1,3 @@
 pub mod install;
+pub mod selfie;
 pub mod switch;
